Refactored routes and topbar example to better paths/url
